+++ bindings/ergo-lib-c/src/rest/api.rs
@@ -2,3 +2,4 @@
 
 pub mod node;
 pub mod node_async;
+pub mod runtime;
