+++ bindings/ergo-lib-ios/Tests/ErgoLibTests/RestNodeApiTests.swift
@@ -7,7 +7,8 @@ final class RestNodeApiTests: XCTestCase {
 
     func testGetInfo() throws {
         let nodeConf = try NodeConf(withAddrString: "213.239.193.208:9053")
-        let nodeInfo = try RestNodeApi.getInfo(nodeConf: nodeConf)
+        let restNodeApi = try RestNodeApi()
+        let nodeInfo = try restNodeApi.getInfo(nodeConf: nodeConf)
         XCTAssert(!nodeInfo.getName().isEmpty)
     }
 }
